refactor: move web3 wrapper in mev_inspect package
